requirements.txt
psycopg2-binary>=2.8.6
PyQt5>=5.15.0
